Add NLPModels to benchmarks
